# openapi.yaml
openapi: 3.0.3
info:
  title: ELS Feedback & Notification API
  version: 3.0.0
servers:
  - url: https://feedback.els.ericterminal.com
paths:
...
  /v1/feedback/issues:
    post:
      summary: 创建反馈工单
      parameters:
        - in: header
          name: X-ELS-Challenge-Id
          required: true
          schema:
            type: string
        - in: header
          name: X-ELS-Timestamp
          required: true
          schema:
            type: string
        - in: header
          name: X-ELS-Signature
          required: true
          schema:
            type: string
        - in: header
          name: X-ELS-PoW-Nonce
          required: true
          schema:
            type: string
        - in: header
          name: X-ELS-PoW-Hash
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  issue_number:
                    type: integer
                  ticket_token:
                    type: string
                  public_url:
                    type: string
                    format: uri
                  status:
                    type: string
        '202':
          description: 已受理但内容被暂时隐藏
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  issue_number:
                    type: integer
                  ticket_token:
                    type: string
                  public_url:
                    type: string
                    format: uri
                  status:
                    type: string
                    enum: [blocked]
                  moderation_blocked:
                    type: boolean
                  moderation_message:
                    type: string
                  archive_id:
                    type: string
        '401':
          description: 签名或 challenge 校验失败
        '429':
          description: 触发限流
  /v1/feedback/issues/{issue_number}:
    get:
      summary: 查询反馈工单状态
      parameters:
        - in: path
          name: issue_number
          required: true
          schema:
            type: integer
        - in: query
          name: ticket_token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
        '403':
          description: ticket_token 无效
  /v1/feedback/issues/{issue_number}/comments:
    post:
      summary: 在反馈工单下发送评论
      parameters:
        - in: path
          name: issue_number
          required: true
          schema:
            type: integer
        - in: query
          name: ticket_token
          required: true
          schema:
            type: string
        - in: header
          name: X-ELS-Challenge-Id
          required: true
          schema:
            type: string
        - in: header
          name: X-ELS-Timestamp
          required: true
          schema:
            type: string
        - in: header
          name: X-ELS-Signature
          required: true
          schema:
            type: string
        - in: header
          name: X-ELS-PoW-Nonce
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
      responses:
        '200':
          description: 评论已发布
        '202':
          description: 评论被暂时隐藏，已发布占位评论
        '403':
          description: ticket_token 无效
components:
  securitySchemes:
    announcementAdminToken:
      type: http
      scheme: bearer
      description: 使用 ANNOUNCEMENT_ADMIN_TOKEN；浏览器管理台使用安全会话 Cookie
...